Refactor workflow chatbot structure and update routing
- Remove the Workflow Chatbot page and its associated components to streamline the tools section.
- Update the Tools home page to redirect the Workflow Chatbot link to a new path.
- Adjust import paths in the Text Analyzer page for consistency.
- Fix SQL policies in Supabase migrations to ensure proper type handling for clerk IDs.

This commit enhances the overall organization of the tools and improves database security policies.
