+++ src/app/tools/chat/components/index.ts
@@ -1,0 +1,7 @@
+export { default as Error_display } from "./error_display";
+export { default as Header } from "./header";
+export { default as Input_area } from "./input_area";
+export { default as Message_area } from "./message_area";
+export { default as New_workflow_modal } from "./new_workflow_modal";
+export { default as Templates_modal } from "./templates_modal";
+export { default as Workflow_panel } from "./workflow_panel";
