+++ src/app/tools/chat/components/workflow_nodes.tsx
